+++ src/email_builder/message.rs
@@ -94,11 +94,11 @@ impl Message {
 }
 
 fn html_template(content: &str) -> String {
-    "<html>
-    <head></head>
-    <body>"
-        .to_string()
-        + content
-        + "</body>
-    </html>"
+    format!(
+        "<html>
+        <head></head>
+        <body>{}</body>
+        </html>",
+        content
+    )
 }